Decompose main.go into per-command files
- config.go: CLIConfig, IPLocation, config load/save/getConfigPath
- util.go: setupClient, writeJSON, formatTime, padOrTruncate, findClassByID
- schedules.go: schedulesCmd and its flag vars
- bookings.go: bookingsCmd, listBookingsCmd, cancelBookingCmd and their vars
- configure.go: configureCmd, configureStudiosCmd, configureTimezoneCmd and vars
- main.go: reduced to rootCmd + main() + init() flag registration
- .gitignore: ignore compiled otf-cli binary
