+++ .gitignore
@@ -2,3 +2,4 @@
 .env
 bin/
 .vscode/
+otf-cli
